+++ eth-rpc-adapter/src/server.ts
@@ -1,52 +1,98 @@
+import http, { ServerOptions } from 'http';
+import WebSocket from 'ws';
+import cors from 'cors';
 import { json as jsonParser } from 'body-parser';
 import connect, { HandleFunction } from 'connect';
-import cors from 'cors';
-import http, { ServerOptions } from 'http';
 import { logger } from './logger';
 import { errorHandler } from './middlewares';
-import { BatchSizeError } from './errors';
-
-import WebSocket from 'ws';
-import { InvalidRequest, MethodNotFound } from './errors';
+import { InvalidRequest, MethodNotFound, BatchSizeError } from './errors';
 import { Router } from './router';
 import { DataDogUtil } from './utils';
-import type { JSONRPCRequest, JSONRPCResponse } from './transports/types';
 
 export interface EthRpcServerOptions extends ServerOptions {
-  middleware: HandleFunction[];
   port: number;
-  cors?: cors.CorsOptions;
   batchSize: number;
+  middleware?: HandleFunction[];
+  cors?: cors.CorsOptions;
+}
+
+export interface JSONRPCRequest {
+  jsonrpc: string;
+  id: string;
+  method: string;
+  params: any[] | Record<string, unknown>;
+}
+
+export interface JSONRPCErrorObject {
+  code: number;
+  message: string;
+  data?: any;
+}
+
+export interface JSONRPCResponse {
+  jsonrpc: string;
+  id: string | null;
+  result?: any;
+  error?: JSONRPCErrorObject;
 }
 
 export default class EthRpcServer {
   private static defaultCorsOptions = { origin: '*' };
   private server: http.Server;
+  private wss: WebSocket.Server;
   private options: EthRpcServerOptions;
-
   public routers: Router[] = [];
 
   constructor(options: EthRpcServerOptions) {
+    this.options = options;
+
     const app = connect();
 
-    const corsOptions = options.cors || EthRpcServer.defaultCorsOptions;
-    this.options = {
-      ...options,
-      middleware: [
-        cors(corsOptions) as HandleFunction,
-        jsonParser({
-          limit: '1mb'
-        }),
-        ...options.middleware
-      ]
-    };
-
-    this.options.middleware.forEach((mw) => app.use(mw));
-
-    app.use(this.httpRouterHandler.bind(this) as HandleFunction);
+    options.middleware?.forEach((mw) => app.use(mw));
+    const corsOptions = options.cors ?? EthRpcServer.defaultCorsOptions;
+    app.use(cors(corsOptions));
+    app.use(jsonParser({ limit: '1mb' }));
+    app.use(this.httpRouterHandler.bind(this));
     app.use(errorHandler);
 
     this.server = http.createServer(app);
+
+    /* ------------------------- wss ---------------------------- */
+    this.wss = new WebSocket.Server({ server: this.server, perMessageDeflate: false });
+
+    this.wss.on('connection', (ws: WebSocket) => {
+      // @ts-ignore
+      ws.isAlive = true;
+
+      ws.on('pong', () => {
+        // @ts-ignore
+        ws.isAlive = true;
+      });
+
+      ws.on('message', (message: string) => this.webSocketRouterHandler(message, ws));
+
+      ws.on('close', () => {
+        ws.removeAllListeners();
+      });
+      ws.on('error', () => {
+        ws.removeAllListeners();
+      });
+    });
+
+    const interval = setInterval(() => {
+      for (const ws of this.wss.clients) {
+        // @ts-ignore
+        if (ws.isAlive === false) return ws.terminate();
+
+        // @ts-ignore
+        ws.isAlive = false;
+        ws.ping();
+      }
+    }, 30000);
+
+    this.wss.on('close', () => {
+      clearInterval(interval);
+    });
   }
 
   public start(): void {
@@ -139,5 +185,46 @@ export default class EthRpcServer {
 
     res.setHeader('Content-Type', 'application/json');
     res.end(JSON.stringify(result));
+  }
+
+  private praseRequest(req: string): any {
+    try {
+      return JSON.parse(req);
+    } catch (e) {
+      return null;
+    }
+  }
+
+  private async webSocketRouterHandler(rawReq: any, ws: WebSocket): Promise<void> {
+    const req = this.praseRequest(rawReq);
+    if (req === null) {
+      const result = {
+        id: null,
+        jsonrpc: '2.0',
+        error: new InvalidRequest().json()
+      };
+      ws.send(JSON.stringify(result));
+      return;
+    }
+
+    let result = null;
+    logger.debug(req, 'WS incoming request');
+
+    if (req instanceof Array) {
+      if (req.length > this.options.batchSize) {
+        result = {
+          jsonrpc: '2.0',
+          error: new BatchSizeError(this.options.batchSize, req.length).json()
+        };
+      } else {
+        result = await Promise.all(req.map((r: JSONRPCRequest) => this.routerHandler(r, ws)));
+      }
+    } else {
+      result = await this.routerHandler(req, ws);
+    }
+
+    logger.debug(result, 'request completed');
+
+    ws.send(JSON.stringify(result));
   }
 }
